+++ lib/relapse/builders.rb
@@ -6,6 +6,6 @@ module Relapse
   end
 end
 
-%w[osx_app source windows_folder windows_folder_from_wrapper windows_installer windows_standalone].each do |builder|
+%w[osx_app source windows_folder windows_folder_from_ruby_dist windows_installer windows_standalone].each do |builder|
   require "relapse/builders/#{builder}"
 end