+++ lib/relapse/project.rb
@@ -104,7 +104,7 @@ module Relapse
 
     # Add a type of output to produce. Must define at least one of these.
     #
-    # @param type [:osx_app, :source, :windows_folder, :windows_folder_from_wrapper, :windows_installer, :windows_standalone]
+    # @param type [:osx_app, :source, :windows_folder, :windows_folder_from_ruby_dist, :windows_installer, :windows_standalone]
     # @return [Project] self
     def add_build(type, &block)
       raise ArgumentError, "Unsupported output type #{type}" unless Builders.has_type? type